scx_layered: fix the gpu matcher
I noticed when the program first started, the taskc->running_at can still be zero
when the kprobe is fired. This will record timestamp as zero.

In the match_one, if the timestamp is zero, we immediately return !must_be_used.
This cause the match to fail. By setting timestamp to now when kprobe is firing
We are always able to match.

Signed-off-by: Patrick Lu <perf.patrick.lu@gmail.com>